Refactor: Dooingle과 Catch에서 NotificationService만 의존하도록 수정
- DooingleService와 CatchService에서 NotificationRepository, SseEmitters 의존 제거
- NotificationService에서 NofiticationRepository와 SseEmitters 의존
