+++ src/main/kotlin/com/dooingle/domain/catch/service/CatchService.kt
@@ -5,11 +5,7 @@ import com.dooingle.domain.catch.dto.CatchResponse
 import com.dooingle.domain.catch.dto.DeleteCatchRequest
 import com.dooingle.domain.catch.repository.CatchRepository
 import com.dooingle.domain.dooingle.repository.DooingleRepository
-import com.dooingle.domain.notification.NotificationRepository
-import com.dooingle.domain.notification.dto.NotificationResponse
-import com.dooingle.domain.notification.model.Notification
-import com.dooingle.domain.notification.model.NotificationType
-import com.dooingle.global.sse.SseEmitters
+import com.dooingle.domain.notification.service.NotificationService
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -18,8 +14,7 @@ import org.springframework.transaction.annotation.Transactional
 class CatchService (
     private val dooingleRepository: DooingleRepository,
     private val catchRepository: CatchRepository,
-    private val notificationRepository: NotificationRepository,
-    private val sseEmitters: SseEmitters
+    private val notificationService: NotificationService
 ){
     // 캐치 생성
     fun addCatch(dooingleId: Long, addCatchRequest: AddCatchRequest): CatchResponse {
@@ -36,17 +31,7 @@ class CatchService (
         dooingle.catch = catch
         catchRepository.save(catch)
 
-        notificationRepository.save(
-            Notification(
-                user = dooingle.owner,
-                notificationType = NotificationType.CATCH,
-                resourceId = dooingleId
-            )
-        ).let {
-            NotificationResponse.from(it)
-        }.let {
-            sseEmitters.sendUserNotification(dooingle.owner.id!!, "${it.message}-${it.cursor}")
-        }
+        notificationService.addCatchNotification(user = dooingle.owner, dooingleId = dooingleId)
 
         return CatchResponse.from(catch)
     }
