+++ src/main/kotlin/com/dooingle/domain/notification/repository/NotificationRepository.kt
@@ -1,4 +1,4 @@
-package com.dooingle.domain.notification
+package com.dooingle.domain.notification.repository
 
 import com.dooingle.domain.notification.model.Notification
 import org.springframework.data.jpa.repository.JpaRepository
